msm: kgsl: Add uintptr_t cast for gpuaddr
gpuaddr in kgsl_memdesc is still an unsigned int. It needs to be
cast to a uintptr_t first and can then be cast to a pointer.

+++ drivers/gpu/msm/kgsl_debugfs.c
@@ -1,4 +1,4 @@
-/* Copyright (c) 2002,2008-2013, The Linux Foundation. All rights reserved.
+/* Copyright (c) 2002,2008-2014, The Linux Foundation. All rights reserved.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2 and
@@ -172,7 +172,7 @@ static void print_mem_entry(struct seq_file *s, struct kgsl_mem_entry *entry)
 	kgsl_get_memory_usage(usage, sizeof(usage), m->flags);
 
 	seq_printf(s, "%pK %pK %8zd %5d %5s %10s %16s %5d\n",
-			(unsigned long *) m->gpuaddr,
+			(unsigned long *)(uintptr_t) m->gpuaddr,
 			(unsigned long *) m->useraddr,
 			m->size, entry->id, flags,
 			memtype_str(entry->memtype), usage, m->sglen);
